ICML-2026/paper-2010-DistMatch/best_code/conformal-pip-requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
brotlipy==0.7.0
cachetools==5.2.0
calmsize==0.1.3
catboost==1.1
click==8.1.3
cloudpickle==2.2.0
cmdstanpy==1.0.7
contourpy==1.0.5
convertdate==2.4.0
cryptography
cycler==0.11.0
Cython==0.29.32
darts==0.22.0
decorator==5.1.1
docker-pycreds==0.4.0
doubt==4.3.0
ephem==4.1.3
et-xmlfile==1.1.0
executing==1.1.1
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.8.2
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.13.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.49.1
hijri-converter==2.2.4
holidays==0.16
hopfield-layers @ git+https://github.com/ml-jku/hopfield-layers@f56f929c95b77a070ae675ea4f56b6d54d36e730
hydra-core==1.2.0
hydra-joblib-launcher==1.2.0
idna
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lightgbm==3.3.3
llvmlite==0.39.1
LunarCalendar==0.0.9
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
multidict==6.0.2
nfoursid==1.0.1
numba==0.56.3
numexpr==2.8.3
numpy
oauthlib==3.2.2
omegaconf==2.2.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
pmdarima==2.0.1
promise==2.3
prompt-toolkit==3.0.31
prophet==1.1.1
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pyDeprecate==0.3.2
Pygments==2.13.0
PyMeeus==0.5.11
pyOpenSSL
pyparsing==3.0.9
PySocks
python-dateutil==2.8.2
pytorch-lightning==1.7.7
pytorch-memlab==0.2.4
pytz==2022.4
PyYAML==6.0
requests
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
sentry-sdk==1.9.10
setproctitle==1.3.2
setuptools-git==1.2
shap==0.41.0
shortuuid==1.0.9
six
slicer==0.0.7
smmap==5.0.0
stack-data==0.5.1
statsforecast==1.1.1
statsmodels==0.13.2
tables==3.7.0
tbats==1.1.1
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.10.0
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.4.0
typing_extensions
ujson==5.5.0
urllib3
wandb==0.13.4
wcwidth==0.2.5
Werkzeug==2.2.2
xarray==2022.10.0
xlrd==2.0.1
yarl==1.8.1
